Updated README, created correct TCP byte stream message delimitation
Now using correct message delimitation with writing the length of the messages
at the beginning instead of relying on timings. This is good as it will now
work if the server is on a different computer.

Updated README, mostly the future expansion section after getting graded
feedback

Moved the phishing image (cake schedule) and its base64 file into the /user/
directory because that makes more sense

Added basickeylogger file created for demo vidoe